fix: 인트로를 위한 레이아웃 분리
