+++ src/app/(with-layout)/card/(components)/badge-status.tsx
