+++ src/app/(with-layout)/card/(components)/pay-notice.tsx
