+++ src/app/(with-layout)/chat/page.tsx
