+++ src/app/(with-layout)/forest/(components)/event-slider-banner.tsx
