+++ src/app/(with-layout)/forest/(components)/ranking.tsx
