+++ src/app/(with-layout)/funding/(_dummy)/detail-data.ts
