+++ src/app/(with-layout)/funding/(components)/ad-slider-banner.tsx
