+++ src/app/(with-layout)/funding/list/[id]/page.tsx
