+++ src/app/(with-layout)/payments/(components)/payments-item.tsx
