+++ src/app/(with-layout)/payments/(components)/payments-list.tsx
